requirements.txt
pillow
flask
redis